Migrate genfrac, lap, smash, op, mod, and katex into their own files in src/functions
Fixes #869
