+++ src/buildHTML.js
@@ -452,147 +452,6 @@ groupTypes.supsub = function(group, options) {
         options);
 };
 
-groupTypes.genfrac = function(group, options) {
-    // Fractions are handled in the TeXbook on pages 444-445, rules 15(a-e).
-    // Figure out what style this fraction should be in based on the
-    // function used
-    let style = options.style;
-    if (group.value.size === "display") {
-        style = Style.DISPLAY;
-    } else if (group.value.size === "text") {
-        style = Style.TEXT;
-    }
-
-    const nstyle = style.fracNum();
-    const dstyle = style.fracDen();
-    let newOptions;
-
-    newOptions = options.havingStyle(nstyle);
-    const numerm = buildGroup(group.value.numer, newOptions, options);
-
-    newOptions = options.havingStyle(dstyle);
-    const denomm = buildGroup(group.value.denom, newOptions, options);
-
-    let rule;
-    let ruleWidth;
-    let ruleSpacing;
-    if (group.value.hasBarLine) {
-        rule = makeLineSpan("frac-line", options);
-        ruleWidth = rule.height;
-        ruleSpacing = rule.height;
-    } else {
-        rule = null;
-        ruleWidth = 0;
-        ruleSpacing = options.fontMetrics().defaultRuleThickness;
-    }
-
-    // Rule 15b
-    let numShift;
-    let clearance;
-    let denomShift;
-    if (style.size === Style.DISPLAY.size) {
-        numShift = options.fontMetrics().num1;
-        if (ruleWidth > 0) {
-            clearance = 3 * ruleSpacing;
-        } else {
-            clearance = 7 * ruleSpacing;
-        }
-        denomShift = options.fontMetrics().denom1;
-    } else {
-        if (ruleWidth > 0) {
-            numShift = options.fontMetrics().num2;
-            clearance = ruleSpacing;
-        } else {
-            numShift = options.fontMetrics().num3;
-            clearance = 3 * ruleSpacing;
-        }
-        denomShift = options.fontMetrics().denom2;
-    }
-
-    let frac;
-    if (ruleWidth === 0) {
-        // Rule 15c
-        const candidateClearance =
-            (numShift - numerm.depth) - (denomm.height - denomShift);
-        if (candidateClearance < clearance) {
-            numShift += 0.5 * (clearance - candidateClearance);
-            denomShift += 0.5 * (clearance - candidateClearance);
-        }
-
-        frac = buildCommon.makeVList({
-            positionType: "individualShift",
-            children: [
-                {type: "elem", elem: denomm, shift: denomShift},
-                {type: "elem", elem: numerm, shift: -numShift},
-            ],
-        }, options);
-    } else {
-        // Rule 15d
-        const axisHeight = options.fontMetrics().axisHeight;
-
-        if ((numShift - numerm.depth) - (axisHeight + 0.5 * ruleWidth) <
-                clearance) {
-            numShift +=
-                clearance - ((numShift - numerm.depth) -
-                             (axisHeight + 0.5 * ruleWidth));
-        }
-
-        if ((axisHeight - 0.5 * ruleWidth) - (denomm.height - denomShift) <
-                clearance) {
-            denomShift +=
-                clearance - ((axisHeight - 0.5 * ruleWidth) -
-                             (denomm.height - denomShift));
-        }
-
-        const midShift = -(axisHeight - 0.5 * ruleWidth);
-
-        frac = buildCommon.makeVList({
-            positionType: "individualShift",
-            children: [
-                {type: "elem", elem: denomm, shift: denomShift},
-                {type: "elem", elem: rule,   shift: midShift},
-                {type: "elem", elem: numerm, shift: -numShift},
-            ],
-        }, options);
-    }
-
-    // Since we manually change the style sometimes (with \dfrac or \tfrac),
-    // account for the possible size change here.
-    newOptions = options.havingStyle(style);
-    frac.height *= newOptions.sizeMultiplier / options.sizeMultiplier;
-    frac.depth *= newOptions.sizeMultiplier / options.sizeMultiplier;
-
-    // Rule 15e
-    let delimSize;
-    if (style.size === Style.DISPLAY.size) {
-        delimSize = options.fontMetrics().delim1;
-    } else {
-        delimSize = options.fontMetrics().delim2;
-    }
-
-    let leftDelim;
-    let rightDelim;
-    if (group.value.leftDelim == null) {
-        leftDelim = makeNullDelimiter(options, ["mopen"]);
-    } else {
-        leftDelim = delimiter.customSizedDelim(
-            group.value.leftDelim, delimSize, true,
-            options.havingStyle(style), group.mode, ["mopen"]);
-    }
-    if (group.value.rightDelim == null) {
-        rightDelim = makeNullDelimiter(options, ["mclose"]);
-    } else {
-        rightDelim = delimiter.customSizedDelim(
-            group.value.rightDelim, delimSize, true,
-            options.havingStyle(style), group.mode, ["mclose"]);
-    }
-
-    return makeSpan(
-        ["mord"].concat(newOptions.sizingClasses(options)),
-        [leftDelim, makeSpan(["mfrac"], [frac]), rightDelim],
-        options);
-};
-
 groupTypes.spacing = function(group, options) {
     if (group.value === "\\ " || group.value === "\\space" ||
         group.value === " " || group.value === "~") {
@@ -615,318 +474,7 @@ groupTypes.spacing = function(group, options) {
     }
 };
 
-groupTypes.lap = function(group, options) {
-    // mathllap, mathrlap, mathclap
-    let inner;
-    if (group.value.alignment === "clap") {
-        // ref: https://www.math.lsu.edu/~aperlis/publications/mathclap/
-        inner = makeSpan([], [buildGroup(group.value.body, options)]);
-        // wrap, since CSS will center a .clap > .inner > span
-        inner = makeSpan(["inner"], [inner], options);
-    } else {
-        inner = makeSpan(
-            ["inner"], [buildGroup(group.value.body, options)]);
-    }
-    const fix = makeSpan(["fix"], []);
-    return makeSpan(
-        ["mord", group.value.alignment], [inner, fix], options);
-};
-
-groupTypes.smash = function(group, options) {
-    const node = makeSpan(["mord"], [buildGroup(group.value.body, options)]);
-
-    if (!group.value.smashHeight && !group.value.smashDepth) {
-        return node;
-    }
-
-    if (group.value.smashHeight) {
-        node.height = 0;
-        // In order to influence makeVList, we have to reset the children.
-        if (node.children) {
-            for (let i = 0; i < node.children.length; i++) {
-                node.children[i].height = 0;
-            }
-        }
-    }
-
-    if (group.value.smashDepth) {
-        node.depth = 0;
-        if (node.children) {
-            for (let i = 0; i < node.children.length; i++) {
-                node.children[i].depth = 0;
-            }
-        }
-    }
-
-    // At this point, we've reset the TeX-like height and depth values.
-    // But the span still has an HTML line height.
-    // makeVList applies "display: table-cell", which prevents the browser
-    // from acting on that line height. So we'll call makeVList now.
-
-    return buildCommon.makeVList({
-        positionType: "firstBaseline",
-        children: [{type: "elem", elem: node}],
-    }, options);
-};
-
-groupTypes.op = function(group, options) {
-    // Operators are handled in the TeXbook pg. 443-444, rule 13(a).
-    let supGroup;
-    let subGroup;
-    let hasLimits = false;
-    if (group.type === "supsub") {
-        // If we have limits, supsub will pass us its group to handle. Pull
-        // out the superscript and subscript and set the group to the op in
-        // its base.
-        supGroup = group.value.sup;
-        subGroup = group.value.sub;
-        group = group.value.base;
-        hasLimits = true;
-    }
-
-    const style = options.style;
-
-    // Most operators have a large successor symbol, but these don't.
-    const noSuccessor = [
-        "\\smallint",
-    ];
-
-    let large = false;
-    if (style.size === Style.DISPLAY.size &&
-        group.value.symbol &&
-        !utils.contains(noSuccessor, group.value.body)) {
-
-        // Most symbol operators get larger in displaystyle (rule 13)
-        large = true;
-    }
-
-    let base;
-    if (group.value.symbol) {
-        // If this is a symbol, create the symbol.
-        const fontName = large ? "Size2-Regular" : "Size1-Regular";
-        base = buildCommon.makeSymbol(
-            group.value.body, fontName, "math", options,
-            ["mop", "op-symbol", large ? "large-op" : "small-op"]);
-    } else if (group.value.value) {
-        // If this is a list, compose that list.
-        const inner = buildExpression(group.value.value, options, true);
-        if (inner.length === 1 && inner[0] instanceof domTree.symbolNode) {
-            base = inner[0];
-            base.classes[0] = "mop"; // replace old mclass
-        } else {
-            base = makeSpan(["mop"], inner, options);
-        }
-    } else {
-        // Otherwise, this is a text operator. Build the text from the
-        // operator's name.
-        // TODO(emily): Add a space in the middle of some of these
-        // operators, like \limsup
-        const output = [];
-        for (let i = 1; i < group.value.body.length; i++) {
-            output.push(buildCommon.mathsym(group.value.body[i], group.mode));
-        }
-        base = makeSpan(["mop"], output, options);
-    }
-
-    // If content of op is a single symbol, shift it vertically.
-    let baseShift = 0;
-    let slant = 0;
-    if (base instanceof domTree.symbolNode) {
-        // Shift the symbol so its center lies on the axis (rule 13). It
-        // appears that our fonts have the centers of the symbols already
-        // almost on the axis, so these numbers are very small. Note we
-        // don't actually apply this here, but instead it is used either in
-        // the vlist creation or separately when there are no limits.
-        baseShift = (base.height - base.depth) / 2 -
-            options.fontMetrics().axisHeight;
-
-        // The slant of the symbol is just its italic correction.
-        slant = base.italic;
-    }
-
-    if (hasLimits) {
-        // IE 8 clips \int if it is in a display: inline-block. We wrap it
-        // in a new span so it is an inline, and works.
-        base = makeSpan([], [base]);
-
-        let supm;
-        let supKern;
-        let subm;
-        let subKern;
-        let newOptions;
-        // We manually have to handle the superscripts and subscripts. This,
-        // aside from the kern calculations, is copied from supsub.
-        if (supGroup) {
-            newOptions = options.havingStyle(style.sup());
-            supm = buildGroup(supGroup, newOptions, options);
-
-            supKern = Math.max(
-                options.fontMetrics().bigOpSpacing1,
-                options.fontMetrics().bigOpSpacing3 - supm.depth);
-        }
-
-        if (subGroup) {
-            newOptions = options.havingStyle(style.sub());
-            subm = buildGroup(subGroup, newOptions, options);
-
-            subKern = Math.max(
-                options.fontMetrics().bigOpSpacing2,
-                options.fontMetrics().bigOpSpacing4 - subm.height);
-        }
-
-        // Build the final group as a vlist of the possible subscript, base,
-        // and possible superscript.
-        let finalGroup;
-        let top;
-        let bottom;
-        if (!supGroup) {
-            top = base.height - baseShift;
-
-            // Shift the limits by the slant of the symbol. Note
-            // that we are supposed to shift the limits by 1/2 of the slant,
-            // but since we are centering the limits adding a full slant of
-            // margin will shift by 1/2 that.
-            finalGroup = buildCommon.makeVList({
-                positionType: "top",
-                positionData: top,
-                children: [
-                    {type: "kern", size: options.fontMetrics().bigOpSpacing5},
-                    {type: "elem", elem: subm, marginLeft: -slant + "em"},
-                    {type: "kern", size: subKern},
-                    {type: "elem", elem: base},
-                ],
-            }, options);
-        } else if (!subGroup) {
-            bottom = base.depth + baseShift;
-
-            finalGroup = buildCommon.makeVList({
-                positionType: "bottom",
-                positionData: bottom,
-                children: [
-                    {type: "elem", elem: base},
-                    {type: "kern", size: supKern},
-                    {type: "elem", elem: supm, marginLeft: slant + "em"},
-                    {type: "kern", size: options.fontMetrics().bigOpSpacing5},
-                ],
-            }, options);
-        } else if (!supGroup && !subGroup) {
-            // This case probably shouldn't occur (this would mean the
-            // supsub was sending us a group with no superscript or
-            // subscript) but be safe.
-            return base;
-        } else {
-            bottom = options.fontMetrics().bigOpSpacing5 +
-                subm.height + subm.depth +
-                subKern +
-                base.depth + baseShift;
-
-            finalGroup = buildCommon.makeVList({
-                positionType: "bottom",
-                positionData: bottom,
-                children: [
-                    {type: "kern", size: options.fontMetrics().bigOpSpacing5},
-                    {type: "elem", elem: subm, marginLeft: -slant + "em"},
-                    {type: "kern", size: subKern},
-                    {type: "elem", elem: base},
-                    {type: "kern", size: supKern},
-                    {type: "elem", elem: supm, marginLeft: slant + "em"},
-                    {type: "kern", size: options.fontMetrics().bigOpSpacing5},
-                ],
-            }, options);
-        }
-
-        return makeSpan(["mop", "op-limits"], [finalGroup], options);
-    } else {
-        if (baseShift) {
-            base.style.position = "relative";
-            base.style.top = baseShift + "em";
-        }
-
-        return base;
-    }
-};
-
-groupTypes.mod = function(group, options) {
-    const inner = [];
-
-    if (group.value.modType === "bmod") {
-        // “\nonscript\mskip-\medmuskip\mkern5mu”
-        if (!options.style.isTight()) {
-            inner.push(makeSpan(
-                ["mspace", "negativemediumspace"], [], options));
-        }
-        inner.push(makeSpan(["mspace", "thickspace"], [], options));
-    } else if (options.style.size === Style.DISPLAY.size) {
-        inner.push(makeSpan(["mspace", "quad"], [], options));
-    } else if (group.value.modType === "mod") {
-        inner.push(makeSpan(["mspace", "twelvemuspace"], [], options));
-    } else {
-        inner.push(makeSpan(["mspace", "eightmuspace"], [], options));
-    }
-
-    if (group.value.modType === "pod" || group.value.modType === "pmod") {
-        inner.push(buildCommon.mathsym("(", group.mode));
-    }
-
-    if (group.value.modType !== "pod") {
-        const modInner = [
-            buildCommon.mathsym("m", group.mode),
-            buildCommon.mathsym("o", group.mode),
-            buildCommon.mathsym("d", group.mode)];
-        if (group.value.modType === "bmod") {
-            inner.push(makeSpan(["mbin"], modInner, options));
-            // “\mkern5mu\nonscript\mskip-\medmuskip”
-            inner.push(makeSpan(["mspace", "thickspace"], [], options));
-            if (!options.style.isTight()) {
-                inner.push(makeSpan(
-                    ["mspace", "negativemediumspace"], [], options));
-            }
-        } else {
-            Array.prototype.push.apply(inner, modInner);
-            inner.push(makeSpan(["mspace", "sixmuspace"], [], options));
-        }
-    }
-
-    if (group.value.value) {
-        Array.prototype.push.apply(inner,
-            buildExpression(group.value.value, options, false));
-    }
-
-    if (group.value.modType === "pod" || group.value.modType === "pmod") {
-        inner.push(buildCommon.mathsym(")", group.mode));
-    }
-
-    return buildCommon.makeFragment(inner);
-};
-
-groupTypes.katex = function(group, options) {
-    // The KaTeX logo. The offsets for the K and a were chosen to look
-    // good, but the offsets for the T, E, and X were taken from the
-    // definition of \TeX in TeX (see TeXbook pg. 356)
-    const k = makeSpan(
-        ["k"], [buildCommon.mathsym("K", group.mode)], options);
-    const a = makeSpan(
-        ["a"], [buildCommon.mathsym("A", group.mode)], options);
-
-    a.height = (a.height + 0.2) * 0.75;
-    a.depth = (a.height - 0.2) * 0.75;
-
-    const t = makeSpan(
-        ["t"], [buildCommon.mathsym("T", group.mode)], options);
-    const e = makeSpan(
-        ["e"], [buildCommon.mathsym("E", group.mode)], options);
-
-    e.height = (e.height - 0.2155);
-    e.depth = (e.depth + 0.2155);
-
-    const x = makeSpan(
-        ["x"], [buildCommon.mathsym("X", group.mode)], options);
-
-    return makeSpan(
-        ["mord", "katex-logo"], [k, a, t, e, x], options);
-};
-
-const makeLineSpan = function(className, options, thickness) {
+export const makeLineSpan = function(className, options, thickness) {
     const line = makeSpan([className], [], options);
     line.height = thickness || options.fontMetrics().defaultRuleThickness;
     line.style.borderBottomWidth = line.height + "em";
