+++ src/functions/operatorname.js
